Refactor workflow evaluation list view and add new components
• Split EmptyState into separate component
• Create TestItem and TestsList components
• Update EvaluationListView with new components
• Add WORKFLOW_EVALUATION_EDIT to constants
• Improve styles and layout of test items
